# server/src/main/resources/static/docs/api/auth-controller-openapi.yaml
openapi: 3.0.3
info:
  title: Dialog Auth API
  description: 인증 및 회원가입 관리를 위한 API
  version: 1.0.0
  contact:
    name: Dialog Team

servers:
  - url: https://api.dialog.com
    description: 프로덕션 서버

tags:
  - name: Auth
    description: 인증 및 회원가입

paths:
  /api/signup:
    post:
      tags:
        - Auth
      summary: 회원가입
      description: |
        OAuth 인증 후 회원가입을 진행합니다.

        **로그인 및 회원가입 프로세스:**
        - GitHub OAuth 또는 Apple 로그인을 통해 인증을 진행합니다
        - 사용자는 등록되지 않은 사용자(TEMP_USER)와 등록된 사용자(USER)로 구분됩니다
        - TEMP_USER: OAuth 인증에 성공했지만 서비스를 처음 이용하여 회원가입이 되지 않은 사용자
        - USER: 서비스에 회원가입한 사용자
        - GitHub: 리다이렉트 시 쿼리 파라미터로 사용자의 `OAUTH_ID`가 전달됩니다
        - Apple: `/api/auth/mobile/apple/login` 호출 후 미가입 사용자인 경우 세션에 OAuth 정보가 저장됩니다
        - 세션 로그인 방식을 사용합니다
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/auth.yaml#/components/schemas/SignupRequest'
            example:
              track: "BACKEND"
              webPushNotification: true
      responses:
        '200':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: './components/schemas/auth.yaml#/components/schemas/SignupResponse'
              example:
                data:
                  userId: 123
        'error-codes':
          description: "[5031]"

  /api/login/check:
    get:
      tags:
        - Auth
      summary: 로그인 상태 확인
      description: 현재 사용자의 로그인 상태를 확인합니다.
      operationId: checkLogin
      parameters:
        - name: JSESSIONID
          in: cookie
          description: 세션 쿠키
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 로그인 상태 확인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: './components/schemas/auth.yaml#/components/schemas/LoginCheckResponse'
              examples:
                loggedIn:
                  summary: 로그인된 상태
                  value:
                    data:
                      isLoggedIn: true
                notLoggedIn:
                  summary: 로그인되지 않은 상태
                  value:
                    data:
                      isLoggedIn: false
        'error-codes':
          description: "[]"

  /api/logout:
    delete:
      tags:
        - Auth
      summary: 로그아웃
      description: 현재 사용자를 로그아웃 처리합니다. 세션 쿠키를 무효화합니다.
      operationId: logout
      parameters:
        - name: JSESSIONID
          in: cookie
          description: 세션 쿠키
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    nullable: true
        'error-codes':
          description: "[]"